Camera Work II & Stairs
Modeled/Imported/Prefab-ed a "Spiral Staircase" asset, and placed it into the scene.
Player can interact with the stairs, which sends them to the upper/lower level of the play area.
When the player does this, the camera follows them and operates its "culling mask", which, when the camera is focused on the "lower" level, will basically make the "upper" level of the play area invisible. It also does this when using the "Tab" key to move between levels (see previous commit).
Later, I would rather make the "upper" level transparent, rather than invisible, possibly by altering the "rendering mode" on the materials of each "upstairs" gameobject to become transparent, and then reduce the alpha to 0(ish). TO DO this, though, I may need to write my own shader, which seems way out of my league, and frustrating as fuck. I'd rather just get on with something simpler for now.