Stairs III & Day/Night Cycle II
Animated "bookshelf" at the top of the stairs so that it opens (automatically for now, when the player gets close).

Added rotation to the day-night cycle light, in an effort to add directional shadows to serve as the game's time-indicator. It looks okay, but not perfect. Might still need to add some other way of telling time.