Day/Night Cycle III & NightManager
Created and animated a text-box for the Computer display, along with several text children and a button.
Created NightManager static class to contain variables used to determine nightly activities (activity type, and difficulty).
Made "dusk" and "dawn" bools in daynightcycle to determine when activities are selected.
The Computer display now shows the player's current state as well as the current nightly activities.