Activity Commencement II
[Committing this now, because I'm probably about to break a lot of shit (See below)]

Moved "CommenceActivities" script to the DialogueManager, which I then renamed GameManager.
Set the time it takes to complete each activity (using a coroutine). Will do further adjustments on each activity's duration later.
The Hero and Sidekick can no longer participate in activities if they are fully fatigued or fully stressed.

[PLAN]
The plan now is to change the activity system a bit. At the moment, 3 activities are set up at night and participation is selected via 3 corresponding toggles. What I want to do is set up just one activity, and have the "activity2" UI elements changed to allow the player/heroes to follow clues picked up during the nightly activities in order to discover and attempt to stop major crimes or "Supervillain Events".
This will work kind of like the board game Clue, where the player has to determine the location of the crime, the actual crime being committed, and the perpetrator. If they can determine these things, they will then have the opportunity to attempt to stop the Villain from succeeding.
This is taking my "computer display" UI in a bit of a different direction, so I will need to delete/alter several things to reflect these changes. At least if I fuck it up, I can return to this point where everything still works.