Event Setup
Added a "ClueMaster" script, which allows the player to collect clues which lead the player to take part in "Events". Each clue points them toward a specific location, target, or type of attack.
Added a "ClueDisplay" UI image, which is parent to several UI Texts, each of which holds the player's clues as they are uncovered. The ClueDisplay is animated using the ComputerDisplay's animator, as it acts in much the same way.