Event Processing
Set up a display (Image>>Texts & Toggles) fo the "Solve attempts".
Created functions to process the player's "solution choices" and determine their correct-ness/success.

I almost fucked something up here with the relevant toggles, trying to get them to only select one at a time, so I'm committing this for peace of mind, because I do NOT want to repeat this shit (which IS what version control is for, yeah?).