Importing Mannequin
Created new "AttackChop" animation for Mannequin model and imported it into project (as a .blend file). Copied once each for the Hero, Enemy, and Civilian, and applied relevant components and scripts for each.
Altered Hero movement using Sebastian Lague's character tutorial (which I also used as reference during the import process.
Removed enemy attacks for now, because it looked stupid and it was pissing me off. Also, since I now have the mannequins working, I shouldn't need the old methods anyway.
Created an "SceneManager - Arena" Gameobject to handle the more scene-based variables and methods (basically spawning and keeping track of enemies and civilians, for now).
Basically, everything should be working as well as, if not better than, it all was before. Also cleaned up some of the scripting.