Player Melee
Added  an empty "HitboxUpper" gameobject to the Hero's hierarchy, which has its own spherecollider.
Altered the "HeroAttack" script to implement the "AttackChop" animation, as well as enabling/disabling the HitboxUpper's collider, which in turn damages any enemy it touches (per their own "EnemyActivity" script).
Added SecondAttack() coroutine to the HeroAttack script, along with relevant variables (and manipulating the Hero's animator component), to allow the Hero to hit with a two-attack combo (of which the second attack is simply the first AttackChop mirrored). Took some adjusting, but it looks good, and doesn't seem to be overlapping states/transitions/whatever.