Map Scene I
Fixed the problems I was having with the editor by re-downloading my Intel Graphics driver).

The Map Scene has been here for the past couple of commits, but this is the first major update focused on it.
-Incorporated player travel between "neighborhoods" (using Input.GetMouseButtonDown and Vector3.Lerp for now).
-Implementing "High-Tier" activities which occur rarely at night and override any "average" activities happening in the neighborhood if the player goes there. It's a lot of small things combined into the whole mechanic, but that's the basics of the work I've been doing tonight, including a lot of revision and such.
-Made the "Day-Night Light" gameobject into a singleton, and some of these new mechanics rely on its current "transition" value.
-If the player is in the "Map Scene" during the daytime (either because time passes, or they return to the map from an activity (though that part is untested), they are prompted to return to the hideout, as they do not fight crime during the day

Lot of fucking work tonight. Not bad for about 4 - 4.5 hours, even as tired as I am. Good stuff. Just gotta keep at it.