Various Improvements
Created a UI element on the Map scene to notify the player of which location is currently experiencing a high-tier activity (as well as the TYPE of activity).
Computer Display updates better (especially when the activity changes as the player is standing at a location).
Improved enemy HP bars, and implemented HP bars for civilians.
All UI canvases/elements scale better with changing screen sizes.
Altered/added nodes and locations to the map.
Altered/added location and attack type clues in the ClueMaster script, and made them consistent with the "event toggles" on the hideout computer clue UI.