Various
Changed "Target Clues" to "Gang Involved Clues" (The point now being to find which of the city's gangs is committing the event. This change has been a long time coming, I just wasn't sure what I wanted to change the "clue type" to, but this will be an improvement).
Adjusted "Event solver" display to reflect changes in event parameters and clues.
Corrected player/enemy/civilian spawning, so they shouldn't spawn inside each others' meshes anymore.
Adjusted enemy attack timing, so they stop "chopping" in sync with each other all the time. Still needs work, but it's an improvement.
Adjusted enemy stat bonuses depending on whether or not the current activity is high-tier (or an event).
Implemented the "DaylightCome" event and added functionality to the scripts which need to pay attention for the event (Not a lot seems to have changed, but maybe it will make more sense in the future. At least for now it's consistent across the multiple scripts "listening" for the event's "signal").
Implemented a "time limit" for Events. Once the player has found the first clue, they SHOULD have a week to solve it before failing (I say SHOULD, because it hasn't been tested, only implemented).
Began implementing "Nightly Overview" UI in the hideout, to keep track of the Hero's exploits during each night and display a break-down of their actions.
Sped up the "Chop" animation in Blender and re-imported, adjusting the player's attack speed to match (I wasn't paying attention to whether it changed for the enemies also, but I think it did. I don't see why it wouldn't).
Created "HitTaken" and "HitTakenDeath" animations and imported them. Implemented enemy death animation into their script. Seems to work quite well.