Combat Animations & Grapple Gun
Created placeholder animations for the following:
-Interacting with objects (at waist level)
-Combat idle loop (when standing still)
-One-handed ranged weapon drawing, firing, and holstering
-Taking damage
-Dying (in place)
-Dashing toward an object hit by the Grappling Hook
-Landing from the "Hook Dash"
-An Overhead Slam when "Hook Dashing" toward an enemy

Added empties to Hero prefab for weapon placement ("thigh holster" and "right-hand-equipped" slots) and parented them to the Hero's armature.

Modeled a basic Grappling Gun and Grappling Hook. Added a trigger collider to the Hook.

Created a "PlayerEquipment" script to hold the player's current loadout (just the gadget for now).

Created a "Gadgets" script with an array for the game's gadgets (only one for now, so it's a pretty short script).

Added a "WeaponCase" to the Hideout, where the player can pick up their gadget(s). Doing so (currently) spawns the Grappling Gun to the player's "Holster" slot.

The player's gadget will also spawn in the holster when entering other scenes.

On the HeroAttack script, incorporated input to initiate the grappling hook (on its own script). Then (to simplify the past 3 days' work), the hook flies either toward the targeted enemy, or just forward, relative to the Hero's rotation.
When the hook hits/triggers a collider, it "anchors" itself and the Hero "dashes" toward it. If the hooked object is an enemy, the Hero performs an overhead smash, damaging the enemy. If the object is not an enemy, the Hero simply lands close by.

The player can also "Cancel" the use of the Grapple Gun after firing, in case they change their mind, I guess.

Holy shit. Lots of work over the past week, and quite a bit of frustration (mostly with the animations and timing), but the new stuff seems to be working okay.
I definitely wouldn't be this far if I hadn't taken a week off from drinking. Good deal.